fix himpunan controller

+++ src/controller/himpunan-controller.js
@@ -5,7 +5,7 @@ const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 import Himpunan from "../model/himpunan.js"
 
-const createHimpunan = async (req, res, next) => {
+const saveHimpunan = async (req, res, next) => {
     const {
         nama_himpunan,
         nama_universitas
@@ -18,27 +18,62 @@ const createHimpunan = async (req, res, next) => {
                 info: 'please upload image'
             });
         }
+
         const {
-            logo_himpunan: [{ path: pathLogoHimpunan }],
             gambar_struktur: [{ path: pathGambarStruktur }],
+            logo_himpunan: [{ path: pathLogoHimpunan }],
         } = req.files
 
-        const newHimpunan = await Himpunan.create({
-            nama_himpunan,
-            nama_universitas,
-            gambar_struktur: pathGambarStruktur,
-            logo_himpunan: pathLogoHimpunan,
-        });
+        const existingHimpunan = await Himpunan.findOne()
+        if (!existingHimpunan) {
+            await Himpunan.create({
+                nama_himpunan,
+                nama_universitas,
+                gambar_struktur: '',
+                logo_himpunan: '',
+            });
+        } else {
+            if (existingHimpunan.gambar_struktur || existingHimpunan.logo_himpunan) {
+                const oldPathGambarStruktur = path.join(__dirname, '../../', existingHimpunan.gambar_struktur)
+                const oldPathLogoHimpunan = path.join(__dirname, '../../', existingHimpunan.logo_himpunan)
 
-        res.status(200).json({
+                fs.unlink(oldPathGambarStruktur, (err) => {
+                    if (err) {
+                        return res.status(400).json({
+                            status: 400,
+                            message: 'failed',
+                            info: 'failed to edit himpunan'
+                        });
+                    }
+                })
+                fs.unlink(oldPathLogoHimpunan, (err) => {
+                    if (err) {
+                        return res.status(400).json({
+                            status: 400,
+                            message: 'failed',
+                            info: 'failed to edit himpunan'
+                        });
+                    }
+                })
+            }
+        }
+
+        await Himpunan.updateOne({}, {
+            $set: {
+                nama_himpunan,
+                nama_universitas,
+                gambar_struktur: pathGambarStruktur,
+                logo_himpunan: pathLogoHimpunan,
+            }
+        })
+        return res.status(200).json({
             status: 200,
             message: 'success',
-            data: newHimpunan
+            data: 'successfully save himpunan'
         })
-
     } catch (error) {
-        console.log(error.message);
-        res.status(500).json({
+        console.log(error.message)
+        return res.status(500).json({
             status: 500,
             message: 'failed',
             info: 'server error'
@@ -64,106 +99,6 @@ const getHimpunan = async (req, res, next) => {
     }
 }
 
-const getHimpunanById = async (req, res, next) => {
-    const { id } = req.params
-    try {
-        const himpunan = await Himpunan.findOne({ _id: id })
-        if (!himpunan) {
-            return res.status(400).json({
-                status: 400,
-                message: 'failed',
-                info: 'himpunan not found'
-            });
-        }
-        res.status(200).json({
-            status: 200,
-            message: 'success',
-            data: himpunan
-        })
-    } catch (error) {
-        console.log(error.message);
-        res.status(500).json({
-            status: 500,
-            message: 'failed',
-            info: 'server error'
-        });
-    }
-}
-
-
-const editHimpunan = async (req, res, next) => {
-    const {
-        nama_himpunan,
-        nama_universitas
-    } = req.body
-    try {
-        if (!req.files.gambar_struktur || !req.files.logo_himpunan) {
-            return res.status(400).json({
-                status: 400,
-                message: 'failed',
-                info: 'please upload image'
-            });
-        }
-
-        const {
-            gambar_struktur: [{ path: pathGambarStruktur }],
-            logo_himpunan: [{ path: pathLogoHimpunan }],
-        } = req.files
-
-        const existingHimpunan = await Himpunan.findOne()
-        if (!existingHimpunan) {
-            return res.status(400).json({
-                status: 400,
-                message: 'failed',
-                info: 'himpunan not found'
-            });
-        }
-
-        const oldPathGambarStruktur = path.join(__dirname, '../../', existingHimpunan.gambar_struktur)
-        const oldPathLogoHimpunan = path.join(__dirname, '../../', existingHimpunan.logo_himpunan)
-
-        fs.unlink(oldPathGambarStruktur, (err) => {
-            if (err) {
-                return res.status(400).json({
-                    status: 400,
-                    message: 'failed',
-                    info: 'failed to edit himpunan'
-                });
-            }
-        })
-        fs.unlink(oldPathLogoHimpunan, (err) => {
-            if (err) {
-                return res.status(400).json({
-                    status: 400,
-                    message: 'failed',
-                    info: 'failed to edit himpunan'
-                });
-            }
-        })
-
-        await Himpunan.updateMany({}, {
-            $set: {
-                nama_himpunan,
-                nama_universitas,
-                gambar_struktur: pathGambarStruktur,
-                logo_himpunan: pathLogoHimpunan,
-            }
-        })
-        return res.status(200).json({
-            status: 200,
-            message: 'success',
-            data: 'successfully edited himpunan'
-        })
-    } catch (error) {
-        console.log(error.message)
-        return res.status(500).json({
-            status: 500,
-            message: 'failed',
-            info: 'server error'
-        });
-    }
-}
-
 const deleteHimpunan = async (req, res, next) => {
     const { id } = req.params
     try {
@@ -176,37 +111,32 @@ const deleteHimpunan = async (req, res, next) => {
             });
         }
 
-        const pathGambarStruktur = path.join(__dirname, '../../', himpunan.gambar_struktur)
-        const pathLogoHimpunan = path.join(__dirname, '../../', himpunan.logo_himpunan)
-        fs.unlink(pathGambarStruktur, (err) => {
-            if (err) {
-                return res.status(400).json({
-                    status: 400,
-                    message: 'failed',
-                    info: 'failed to deleted himpunan'
-                });
-            }
-        })
-        fs.unlink(pathLogoHimpunan, (err) => {
-            if (err) {
-                return res.status(400).json({
-                    status: 400,
-                    message: 'failed',
-                    info: 'failed to deleted himpunan'
-                });
-            }
-        })
+        if (himpunan.gambar_struktur || himpunan.logo_himpunan) {
+            const pathGambarStruktur = path.join(__dirname, '../../', himpunan.gambar_struktur)
+            const pathLogoHimpunan = path.join(__dirname, '../../', himpunan.logo_himpunan)
+            fs.unlink(pathGambarStruktur, (err) => {
+                if (err) {
+                    return res.status(400).json({
+                        status: 400,
+                        message: 'failed',
+                        info: 'failed to deleted himpunan'
+                    });
+                }
+            })
+            fs.unlink(pathLogoHimpunan, (err) => {
+                if (err) {
+                    return res.status(400).json({
+                        status: 400,
+                        message: 'failed',
+                        info: 'failed to deleted himpunan'
+                    });
+                }
+            })
+        }
 
-        await Himpunan.updateMany({}, {
-            $set: {
-                nama_himpunan: '',
-                nama_universitas: '',
-                gambar_struktur: '',
-                logo_himpunan: '',
-            }
-        })
+        await Himpunan.deleteOne({ _id: id })
 
-        res.status(200).json({
+        return res.status(200).json({
             status: 200,
             message: 'success',
             data: 'successfully deleted himpunan'
@@ -222,10 +152,8 @@ const deleteHimpunan = async (req, res, next) => {
 }
 
 const himpunanController = {
-    createHimpunan,
+    saveHimpunan,
     getHimpunan,
-    getHimpunanById,
-    editHimpunan,
     deleteHimpunan,
 }
